refacto code
