+++ src/components/study/rights/NewStudyRightModal.tsx
@@ -1,4 +1,4 @@
-import Modal from '@/components/base/Modal'
+import Modal from '@/components/modals/Modal'
 import { useTranslations } from 'next-intl'
 
 interface Props {
